# backend/menu-api/src/models/MenuItem.mjs
class MenuItem {
  constructor(id, name, description, price, type) {
    this.id = id;
    this.name = name;
    this.description = description;
    this.price = price;
    this.type = type;
  }
}

export { MenuItem };