# backend/openapi.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Swagger Menu
  description: A sample API that uses a restaurant menu
  termsOfService: http://swagger.io/terms/
  contact:
    name: Swagger API Team
    email: apiteam@swagger.io
    url: http://swagger.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://petstore.swagger.io/api
paths:
...
  /menuitem/{id}:
    get:
      tags:
        - Menu Items
      description: Returns a menu item based on a single ID
      operationId: findMenuItemById
      parameters:
        - name: id
          in: path
          description: ID of menu item to fetch
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: menu item response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItem'
              examples:
                menuItem:
                  summary: A Menu Item
                  value:
                    id: 1
                    name: Pandebonos
                    price: 8000
                    description: 6 Pandebonos pequeños.
                    type: 1
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Menu Items
      description: Returns a menu item updated
      operationId: updateMenuItemById
      requestBody:
        description: MenuItem to add to the menu
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuItem'
            examples:
              newMenuItem:
                summary: A menu item to update
                value:
                  id: 1
                  name: Pandebonos
                  price: 8000
                  description: 6 Pandebonos pequeños.
                  type: 1
      parameters:
        - name: id
          in: path
          description: ID of menu item to fetch
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: menu item response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItem'
              examples:
                menuItem:
                  summary: A Menu Item
                  value:
                    id: 1
                    name: Pandebonos
                    price: 8000
                    description: 6 Pandebonos pequeños.
                    type: 1
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Menu Items
      description: deletes a single menu item based on the ID supplied
      operationId: deleteMenuItem
      parameters:
        - name: id
          in: path
          description: ID of menu item to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: menu item deleted
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    MenuItem:
      type: object
      required:
      - id
      - name
      - price
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: integer
          format: int64
        type:
          $ref: "#/components/schemas/MenuItemType"
    MenuItemType:
      type: string
      enum:
        - Entrada
        - PlatoFuerte
        - BebidasPostres
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string